% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{and2abbr}
\alias{and2abbr}
\title{And (\code{&}) to abbreviation (\code{'and'})}
\usage{
and2abbr(x)
}
\arguments{
\item{x}{string}
}
\value{
abbreviated and symbols
}
\description{
And (\code{&}) to abbreviation (\code{'and'})
}
\keyword{internal}
