% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{asterisk2abbr}
\alias{asterisk2abbr}
\title{Asterisk (\code{*}) to abbreviation (\code{'ast'})}
\usage{
asterisk2abbr(x)
}
\arguments{
\item{x}{string}
}
\value{
abbreviated asterisk symbols
}
\description{
Asterisk (\code{*}) to abbreviation (\code{'ast'})
}
\keyword{internal}
