% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_name.R
\name{at2abbr}
\alias{at2abbr}
\title{At (\code{@}) to abbreviation (\code{'at'})}
\usage{
at2abbr(x)
}
\arguments{
\item{x}{string}
}
\value{
abbreviated at symbols
}
\description{
At (\code{@}) to abbreviation (\code{'at'})
}
