% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_name.R
\name{bang2abbr}
\alias{bang2abbr}
\title{Bang (\code{!}) to abbreviation (\code{'bang'})}
\usage{
bang2abbr(x)
}
\arguments{
\item{x}{string}
}
\value{
abbreviated bang symbols
}
\description{
Bang (\code{!}) to abbreviation (\code{'bang'})
}
