% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{caret2abbr}
\alias{caret2abbr}
\title{Caret (hat) (\code{^}) to abbreviation (\code{'hat'})}
\usage{
caret2abbr(x)
}
\arguments{
\item{x}{string}
}
\value{
abbreviated caret symbols
}
\description{
Caret (hat) (\code{^}) to abbreviation (\code{'hat'})
}
\keyword{internal}
