% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{check_chr}
\alias{check_chr}
\title{Check R object name is a character string}
\usage{
check_chr(x)
}
\arguments{
\item{x}{string}
}
\value{
character input
}
\description{
Check R object name is a character string
}
\keyword{internal}
