% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{check_file_extensions}
\alias{check_file_extensions}
\title{Check if name has file extension}
\usage{
check_file_extensions(x)
}
\arguments{
\item{x}{character string}
}
\value{
logical (\code{TRUE} = has file extension, \code{FALSE} = no file extension)
}
\description{
Check if name has file extension
}
\examples{
check_file_extensions("file name with extension.ext")
check_file_extensions("long file name without extension")
}
\keyword{internal}
