% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{check_num_prefix}
\alias{check_num_prefix}
\title{Check R object name for numeric prefix}
\usage{
check_num_prefix(input)
}
\arguments{
\item{x}{string}
}
\value{
character string with numeric values moved to end
}
\description{
Check R object name for numeric prefix
}
\keyword{internal}
