% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{convert_symbols}
\alias{convert_symbols}
\title{Covert symbols to strings}
\usage{
convert_symbols(x)
}
\arguments{
\item{x}{string}
}
\value{
selected converted symbols
}
\description{
Covert symbols to strings
}
\keyword{internal}
