% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{date_formats}
\alias{date_formats}
\alias{date_patterns}
\title{Example date formats and regular expressions}
\format{
\code{date_formats}: character vector (16). \code{date_patterns}: named list (7).

An object of class \code{list} of length 7.
}
\usage{
date_formats

date_patterns
}
\description{
\code{date_formats} is a character vector of common date formats and
\code{date_patterns} is a named list of regular expressions for matching them.
}
\keyword{internal}
