% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{dollar2abbr}
\alias{dollar2abbr}
\title{Dollar (\code{$}) to abbreviation (\code{'dollar'})}
\usage{
dollar2abbr(x)
}
\arguments{
\item{x}{string}
}
\value{
abbreviated dollar sign symbols
}
\description{
Dollar (\code{$}) to abbreviation (\code{'dollar'})
}
\keyword{internal}
