% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{emdash2abbr}
\alias{emdash2abbr}
\title{Em dash (\verb{—}) to abbreviation (\code{'emdash'})}
\usage{
emdash2abbr(x)
}
\arguments{
\item{x}{string}
}
\value{
abbreviated em dash symbols
}
\description{
Em dash (\verb{—}) to abbreviation (\code{'emdash'})
}
\keyword{internal}
