% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{equal2abbr}
\alias{equal2abbr}
\title{Equals (\code{=}) to abbreviation (\code{'equals'})}
\usage{
equal2abbr(x)
}
\arguments{
\item{x}{string}
}
\value{
abbreviated equals symbols
}
\description{
Equals (\code{=}) to abbreviation (\code{'equals'})
}
\keyword{internal}
