% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{exclamation2abbr}
\alias{exclamation2abbr}
\title{Exclamation (\code{!}) to abbreviation (\code{'bang'})}
\usage{
exclamation2abbr(x)
}
\arguments{
\item{x}{string}
}
\value{
abbreviated exclamation symbols
}
\description{
Exclamation (\code{!}) to abbreviation (\code{'bang'})
}
\keyword{internal}
