% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_create.R
\name{ger_create}
\alias{ger_create}
\title{Create a new good enough R project}
\usage{
ger_create(path, open = interactive())
}
\arguments{
\item{path}{path to the new project folder (must not exist)}

\item{open}{logical, open the new project?}
}
\value{
\code{path} (invisibly)
}
\description{
Create a new 'good enough' project folder with all the gerp files and
folders (\code{\link[=ger_setup]{ger_setup()}}, \code{\link[=ger_code]{ger_code()}}, \code{\link[=ger_data]{ger_data()}}, \code{\link[=ger_dev]{ger_dev()}},
\code{\link[=ger_report]{ger_report()}}), a \code{README.Rmd}, and an \code{.Rproj} file. The project is
opened in RStudio or Positron when \code{open = TRUE}.
}
\examples{
tmp <- file.path(tempdir(), "my-project")
ger_create(tmp, open = FALSE)
}
