% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_data.R
\name{ger_data}
\alias{ger_data}
\title{Good enough data folders}
\usage{
ger_data(path = ".")
}
\arguments{
\item{path}{path to project folder}
}
\value{
\code{path} (invisibly)
}
\description{
Create the data folders and a \code{data.md} file:

\describe{
\item{data-raw/}{code used to import, download and include data in your project}
\item{data/}{processed, intermediate, or otherwise altered data in your project}
\item{inst/extdata/}{external data used for examples or testing}
\item{data.md}{document your data here and in \code{R/data.R}}
}
}
\examples{
tmp <- file.path(tempdir(), "data-example")
ger_data(tmp)
}
