% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_go.R
\name{ger_go}
\alias{ger_go}
\title{NEW gerp project!}
\usage{
ger_go(
  path,
  rstudio = rstudioapi::isAvailable(),
  open = rlang::is_interactive()
)
}
\arguments{
\item{path}{path to new \code{gerp} project}

\item{rstudio}{check if RStudio is running}

\item{open}{is R running interactively?}
}
\value{
gerp project
}
\description{
Create a new good enough R project
}
\section{Project functions}{


The following functions are imported from \href{https://github.com/r-lib/usethis/blob/0be0406d14d9816fd7f2457db4f7deaa958ac441/R/proj.R}{usethis}
\subsection{Create project}{
\itemize{
\item \code{proj}: a new environment with \code{new.env(parent = emptyenv())}
}
}

\subsection{\code{get_()}, \code{active()}, \code{set()} projects}{
\itemize{
\item \code{proj_get_}: The current project (\code{proj_get_ <- function() proj$cur})
\item \code{proj_active()}: is function created with \code{!is.null(proj_get_())}
\item \code{proj_set_()}: set \code{old} project path

\if{html}{\out{<div class="sourceCode">}}\preformatted{proj_set_ <- function(path) \{
    old <- proj$cur
    proj$cur <- path
    invisible(old)
\}
}\if{html}{\out{</div>}}
\item \code{proj_get()}: get the current project

\if{html}{\out{<div class="sourceCode">}}\preformatted{proj_get <- function() \{
# Called for first time so try working directory
  if (!proj_active()) \{
    usethis::proj_set(".")
  \}
  proj_get_()
\}
}\if{html}{\out{</div>}}
}

All of these result in the response from \code{usethis::proj_set(".")}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{✔ Setting active project to '/path/to/current/project'
}\if{html}{\out{</div>}}
}
}

\section{UI functions}{

\subsection{UI function (from \href{https://github.com/r-lib/usethis/blob/0be0406d14d9816fd7f2457db4f7deaa958ac441/R/ui.R}{usethis/R/ui.R})}{
}
}

\examples{
# ger_go("path/to/folder")
}
