% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{ger_path_lkp}
\alias{ger_path_lkp}
\title{Get look-up file/folder tree path}
\usage{
ger_path_lkp(x)
}
\arguments{
\item{x}{file or folder to search for}
}
\value{
file tree(s)
}
\description{
Get look-up file/folder tree path
}
\examples{
gerp::ger_path_lkp("inst")
gerp::ger_path_lkp("README.Rmd")
}
