% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_paths.R
\name{ger_paths}
\alias{ger_paths}
\title{Get good enough file paths}
\usage{
ger_paths(..., ext = NULL, rel = TRUE)
}
\arguments{
\item{...}{arguments passed to `dir()`}

\item{ext}{optional file extension (with or without the dot (`.`))}

\item{rel}{logical, relative or absolute path (default is relative)}
}
\value{
file path(s) that are either normalized (`rel = FALSE`) or relative.
}
\description{
Get good enough file paths
}
\examples{
ger_paths(".")
}
