% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_proj.R
\name{ger_proj}
\alias{ger_proj}
\title{Quickly create new gerp project}
\usage{
ger_proj()
}
\value{
downloaded \code{gerp} project (named \code{new_project})
}
\description{
Wrapper around \code{usethis::use_course()} (for downloading a new \code{gerp} project)
}
\examples{
# ger_proj()
}
