% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_ref.R
\name{ger_ref}
\alias{ger_ref}
\title{good enough R references}
\usage{
ger_ref(show_file = TRUE)
}
\arguments{
\item{show_file}{open changelog.md file?}
}
\description{
Create ref folder with ref/notebook.Rmd, ref/manuscript.Rmd,
and changelog.md files.
\describe{
\item{ref/notebook.Rmd}{An R Markdown notebook (i.e., lab notebook).}
\item{ref/manuscript.Rmd}{A more polished R Markdown file for reports or manuscripts.}
\item{changelog.md}{Manually document changes to the files or folders in your project.}
}
}
