% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.R
\name{ger_sect}
\alias{ger_sect}
\title{Insert a fun R script section header \verb{<(+_+)>}}
\usage{
ger_sect(name)
}
\arguments{
\item{name}{section name}
}
\value{
the section header text (invisibly)
}
\description{
Insert a section header at the cursor in RStudio or Positron (similar to
\code{Cmd/Ctrl + Shift + R}).
}
\examples{
# with RStudio or Positron running:
# ger_sect("import")
}
