% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_batch_rename.R
\name{get_clean_file_names}
\alias{get_clean_file_names}
\title{Clean \code{file_name} column (utility)}
\usage{
get_clean_file_names(df, column)
}
\arguments{
\item{df}{data.frame or tibble}

\item{column}{file name column (without extension)}
}
\value{
tibble with clean file names
}
\description{
Clean \code{file_name} column (utility)
}
