% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_batch_rename.R
\name{get_date_prefix}
\alias{get_date_prefix}
\title{Create \code{date_prefix} column (utility)}
\usage{
get_date_prefix(files, prefix = "birth")
}
\arguments{
\item{files}{data.frame or tibble from \code{get_files_tbl()} function}

\item{prefix}{date prefix (\code{"modification"}, \code{"access"}, \code{"change"}, or
\code{"birth"})}
}
\value{
tibble with date prefix column
}
\description{
Create \code{date_prefix} column (utility)
}
