% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_batch_rename.R
\name{get_files_tbl}
\alias{get_files_tbl}
\title{Create \code{files_tbl} table}
\usage{
get_files_tbl(path, ...)
}
\arguments{
\item{path}{path to files to be renamed}

\item{...}{additional args passed to \code{fs::dir_info()}}
}
\value{
tibble with file and date info
}
\description{
Create \code{files_tbl} table
}
