% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_dev.R
\name{grp_dev}
\alias{grp_dev}
\title{Good enough R development folder}
\usage{
grp_dev(path = ".")
}
\arguments{
\item{path}{path to project folder}
}
\value{
\code{path} (invisibly)
}
\description{
Create a \verb{dev/} folder with \code{dev/notebook.Rmd} (a running lab notebook).
}
\examples{
tmp <- file.path(tempdir(), "dev-example")
grp_dev(tmp)
}
