% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{grp_fname}
\alias{grp_fname}
\title{Good enough file name}
\usage{
grp_fname(x, date = Sys.Date(), clip = TRUE)
}
\arguments{
\item{x}{character vector of file names}

\item{date}{date prefix (\code{NULL} for no prefix)}

\item{clip}{logical, copy the name to the clipboard (single names only)?}
}
\value{
a good enough file name (invisibly if copied to the clipboard)
}
\description{
Convert a string into a lowercase, hyphenated file name with a date
prefix (\code{YYYY-MM-DD_}). File extensions are kept.
}
\examples{
grp_fname("November profits (monday)-Carl's copy", clip = FALSE)
grp_fname("\%file & with @_gArbage NAME.txt", date = NULL, clip = FALSE)
}
