% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{grp_fpath}
\alias{grp_fpath}
\title{Get the path of the current file}
\usage{
grp_fpath(tree = FALSE)
}
\arguments{
\item{tree}{logical, print a folder tree?}
}
\value{
file path (invisibly if \code{tree = TRUE})
}
\description{
Return the path to the file open in the RStudio or Positron editor.
}
\examples{
# with RStudio or Positron running:
# grp_fpath()
}
