% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.R
\name{grp_headr}
\alias{grp_headr}
\title{Insert a 'good enough' R script header}
\usage{
grp_headr()
}
\value{
the header text (invisibly)
}
\description{
Insert a script header at the cursor in RStudio or Positron. The header
includes sections for what the code creates, who the author is, the
license, and the version.
}
\examples{
# with RStudio or Positron running:
# grp_headr()
}
