% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{grp_lkp_path}
\alias{grp_lkp_path}
\title{Look up a file or folder in your project}
\usage{
grp_lkp_path(x, path = grp_root())
}
\arguments{
\item{x}{file or folder name to search for (e.g., \code{"README.Rmd"})}

\item{path}{path to search (defaults to the project root)}
}
\value{
matching path(s) (invisibly)
}
\description{
Search the project (from \code{\link[=grp_root]{grp_root()}}) for files or folders named \code{x} and
print a folder tree for each match.
}
\examples{
# grp_lkp_path("README.Rmd")
# grp_lkp_path("inst")
}
