% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{grp_name}
\alias{grp_name}
\title{Good enough R object name}
\usage{
grp_name(x, abbr = FALSE, clip = TRUE)
}
\arguments{
\item{x}{character vector}

\item{abbr}{logical, abbreviate long names?}

\item{clip}{logical, copy the name to the clipboard (single names only)?}
}
\value{
a good enough R object name (invisibly if copied to the clipboard)
}
\description{
Convert a string into a \code{snake_case} R object name. Symbols are replaced
with abbreviations (e.g., \verb{\%} becomes \code{perc}), leading numbers are moved
to the end of the name, and strings without letters become \code{"num_var"}.
}
\examples{
grp_name("2022-10-12-Alpha-20\%", clip = FALSE)
grp_name("Alpha & Beta", clip = FALSE)
}
