% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{grp_path}
\alias{grp_path}
\title{Get file path(s)}
\usage{
grp_path(path = ".", type = c("abs", "rel"), tree = FALSE)
}
\arguments{
\item{path}{file or folder path}

\item{type}{absolute (\code{"abs"}) or relative (\code{"rel"}) path}

\item{tree}{logical, print a folder tree?}
}
\value{
path(s) (invisibly if \code{tree = TRUE})
}
\description{
Use \code{grp_path()} to view folder and file paths. Absolute paths are the
exact location of a file or folder in the file system (they start with the
root folder). Relative paths start from the current working directory.
}
\examples{
grp_path(".")
grp_path(tempdir(), type = "rel")
}
