% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_report.R
\name{grp_report}
\alias{grp_report}
\title{Good enough R report folder}
\usage{
grp_report(path = ".")
}
\arguments{
\item{path}{path to project folder}
}
\value{
\code{path} (invisibly)
}
\description{
Create a \verb{report/} folder with \code{report/manuscript.Rmd} (a more polished R
Markdown file for reports or manuscripts).
}
\examples{
tmp <- file.path(tempdir(), "report-example")
grp_report(tmp)
}
