% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_setup.R
\name{grp_setup}
\alias{grp_setup}
\title{Good enough R setup files}
\usage{
grp_setup(path = ".")
}
\arguments{
\item{path}{path to project folder}
}
\value{
\code{path} (invisibly)
}
\description{
Create start-up files for a project (\code{changelog.md}, \code{CITATION},
\code{requirements.md}, and \code{LICENSE}). Existing files are not overwritten.

\describe{
\item{changelog.md}{Manually document changes to the files or folders in your project.}
\item{CITATION}{Information and example of how to cite your project.}
\item{requirements.md}{Advice on how to manually list the requirements for your project.}
\item{LICENSE}{The \href{https://creativecommons.org/publicdomain/zero/1.0/}{CC0 1.0} license.}
}
}
\examples{
tmp <- file.path(tempdir(), "setup-example")
grp_setup(tmp)
}
