% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{hyphen2abbr}
\alias{hyphen2abbr}
\title{Hyphen (\code{-}) to abbreviation (\code{'hyphen'})}
\usage{
hyphen2abbr(x)
}
\arguments{
\item{x}{string}
}
\value{
abbreviated hyphen symbols
}
\description{
Hyphen (\code{-}) to abbreviation (\code{'hyphen'})
}
\keyword{internal}
