% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{num2abbr}
\alias{num2abbr}
\title{Number symbol (\verb{#}) to abbreviation (\code{'num'})}
\usage{
num2abbr(x)
}
\arguments{
\item{x}{string}
}
\value{
abbreviated plus symbols
}
\description{
Number symbol (\verb{#}) to abbreviation (\code{'num'})
}
\keyword{internal}
