% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_name.R
\name{perc2abbr}
\alias{perc2abbr}
\title{Percents (\verb{\%}) to abbreviation (\code{'perc'})}
\usage{
perc2abbr(x)
}
\arguments{
\item{x}{string}
}
\value{
abbreviated percent symbols
}
\description{
Percents (\verb{\%}) to abbreviation (\code{'perc'})
}
