% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{percents2abbr}
\alias{percents2abbr}
\title{Percents (\verb{\%}) to abbreviation (\code{'perc'})}
\usage{
percents2abbr(x)
}
\arguments{
\item{x}{string}
}
\value{
abbreviated percent symbols
}
\description{
Percents (\verb{\%}) to abbreviation (\code{'perc'})
}
\keyword{internal}
