% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{period2abbr}
\alias{period2abbr}
\title{Period (\code{.}) to abbreviation (\code{'dot'})}
\usage{
period2abbr(x)
}
\arguments{
\item{x}{string}
}
\value{
abbreviated period symbols
}
\description{
Period (\code{.}) to abbreviation (\code{'dot'})
}
\keyword{internal}
