% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_name.R
\name{plus2abbr}
\alias{plus2abbr}
\title{Pluses (\code{+}) to abbreviation (\code{'plus'})}
\usage{
plus2abbr(x)
}
\arguments{
\item{x}{string}
}
\value{
abbreviated plus symbols
}
\description{
Pluses (\code{+}) to abbreviation (\code{'plus'})
}
