% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_batch_rename.R
\name{rename_all_files}
\alias{rename_all_files}
\title{Rename all files (utility)}
\usage{
rename_all_files(old, new)
}
\arguments{
\item{old}{previous names}

\item{new}{new names}
}
\value{
quietly returns the output from \code{purrr::walk2}
}
\description{
Rename all files (utility)
}
