% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{symb2abbr}
\alias{symb2abbr}
\title{Symbol to abbreviations}
\usage{
symb2abbr(input)
}
\arguments{
\item{input}{string}
}
\value{
abbreviated symbols
}
\description{
Symbol to abbreviations
}
\keyword{internal}
