% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{tilde2abbr}
\alias{tilde2abbr}
\title{Tilde symbol (\verb{#}) to abbreviation (\code{'tilde'})}
\usage{
tilde2abbr(x)
}
\arguments{
\item{x}{string}
}
\value{
abbreviated tilde symbols
}
\description{
Tilde symbol (\verb{#}) to abbreviation (\code{'tilde'})
}
\keyword{internal}
