% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_batch_rename.R
\name{ger_batch_rename}
\alias{ger_batch_rename}
\title{Rename all files in a folder}
\usage{
ger_batch_rename(path, prefix)
}
\arguments{
\item{path}{path to folder}

\item{prefix}{date prefix (\code{"modification"}, \code{"access"}, \code{"change"}, or
\code{"birth"})}
}
\value{
returns the output from \code{purrr::walk2}
}
\description{
Rename all files in a folder
}
