% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_code.R
\name{ger_code}
\alias{ger_code}
\title{good enuff code files}
\usage{
ger_code(folder_name = NULL, header = TRUE)
}
\arguments{
\item{folder_name}{name of folder for code files (string)}

\item{header}{include roxygen2 headers?}
}
\description{
Create the code folder for the import, tidy, wrangle, visualize,
    model, and data .R files.
}
\details{
If header is `FALSE`, code files are created with a standard script
   header.
}
