% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_create.R
\name{ger_create}
\alias{ger_create}
\title{Create a new good enough R project}
\usage{
ger_create(folder, name)
}
\arguments{
\item{folder}{the parent folder of your project}

\item{name}{name of your project}
}
\description{
Create and open a new 'good enough' project.
}
\details{
this function will open a new RStudio project. The \code{folder} is
the parent folder that will contain your new project, and the \code{name} will
be the name of your project sub-folder.
To navigate folders on your computer, I like to use \code{ger_path()} function.
If you start with \code{ger_path("/")}, you should be able to find all the
folders you have permissions to (probably best to start with \code{"/Users"}).
}
\examples{
# not run
# ger_create(folder = tempdir(), "my project 01")
}
