% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_data.R
\name{ger_data}
\alias{ger_data}
\title{good enough data folders}
\usage{
ger_data(folder_name = NULL)
}
\description{
Create data folders, inst/extdata folder, and
data.md file
}
\details{
ger_data() will create the following files and folders:
\describe{
\item{data-raw/}{should contain any code used to import, download and include data in your project}
\item{data/}{should contain processed, intermediate, or otherwise altered data in your project}
\item{inst/exdata/}{should have any external data used for examples or testing}
\item{data.md}{Document your data here and in code/data.R}
}
}
\examples{
# in new project
# ger_data()
}
