% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_dev.R
\name{ger_dev}
\alias{ger_dev}
\title{good enough R development folder}
\usage{
ger_dev(folder_name = NULL)
}
\description{
Create \verb{dev/} folder with \code{dev/notebook.Rmd} and \code{changelog.md} files.

\describe{
\item{dev/notebook.Rmd}{An R Markdown notebook (i.e., lab notebook).}
}
}
