% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{ger_fname}
\alias{ger_fname}
\title{Good enough file name}
\usage{
ger_fname(x)
}
\arguments{
\item{x}{string}
}
\value{
good enough file name
}
\description{
Good enough file name
}
\examples{
ger_fname(x = "November profits (monday)-Carl's copy")
ger_fname(x = "\%file & with @_gArbage NAME.txt")
}
