% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{ger_fpath}
\alias{ger_fpath}
\title{Get current file path}
\usage{
ger_fpath(tree = FALSE)
}
\arguments{
\item{tree}{logical, return a folder tree?}
}
\value{
current file path or current file path + folder tree
}
\description{
Get current file path
}
\examples{
# ger_fpath()
}
