% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_fun_headr.R
\name{ger_fun_headr}
\alias{ger_fun_headr}
\title{Fun script header <(+_+)>}
\usage{
ger_fun_headr(name)
}
\arguments{
\item{name}{}
}
\description{
The fun_headr function is similar to the hot keys
'cmd + shift + R' for creating a section header.
}
\examples{
# use this to create a fun section header in any .R file
# with RStudio running...
# geR2::fun_headr("import")

}
