% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ger_headr}
\alias{ger_headr}
\title{Insert 'good enough' R script header}
\usage{
ger_headr()
}
\description{
The `ger_headr()` function can be used in a new R script
    to create a header. The header includes sections for:
* what the code creates

* who the author is

* an MIT License

* Version
}
\examples{
# use this to create a script header (with RStudio open)
# ger_headr("import")
}
