% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.R
\name{ger_headr}
\alias{ger_headr}
\title{Insert 'good enough' R script header}
\usage{
ger_headr()
}
\description{
The \code{ger_headr()} function can be used in a new R script
to create a header. The header includes sections for:
\itemize{
\item what the code creates
\item who the author is
\item an MIT License
\item Version
}
}
\examples{
# use this to create a script header (with RStudio open)
# ger_headr("import")
}
