% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_headr.R
\name{ger_headr}
\alias{ger_headr}
\title{Insert good enuff R script header}
\usage{
ger_headr()
}
\description{
The ger_headr() function can be used in a new R script
    to create a header. The header includes sections for:
- what the code creates
- who the author is
- MIT License
- Version
}
\examples{
# use this to create a script header (with RStudio open)
# ger_headr("import")

}
