% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{ger_lkp_path}
\alias{ger_lkp_path}
\title{Get look-up file/folder tree path}
\usage{
ger_lkp_path(x)
}
\arguments{
\item{x}{file or folder to search for}
}
\value{
file tree(s)
}
\description{
Get look-up file/folder tree path
}
\examples{
gerp::ger_lkp_path("inst")
gerp::ger_lkp_path("README.Rmd")
}
