% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_name.R
\name{ger_name}
\alias{ger_name}
\title{Good enough name}
\usage{
ger_name(x)
}
\arguments{
\item{x}{string}
}
\value{
good enough name (for an R object)
}
\description{
Good enough name
}
\examples{
ger_name("2022-10-12-Alpha-20\%")
ger_name("Alpha & Beta")
}
