% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_paths.R
\name{ger_path}
\alias{ger_path}
\title{Get good enough file path(s)}
\usage{
ger_path(..., ext = NULL, rel = TRUE)
}
\arguments{
\item{...}{arguments passed to \code{dir()}}

\item{ext}{optional file extension (with or without the dot (\code{.}))}

\item{rel}{logical, relative or absolute path (default is relative)}
}
\value{
file path(s) that are either normalized (\code{rel = FALSE}) or relative.
}
\description{
Use \code{ger_path()} to view folder and file paths.
\describe{
\item{ext}{An optional file extension argument (i.e., \code{".csv"})}
\item{rel}{Return the relative or absolute path. The default is a relative path, which is the location of a file or folder \strong{relative to the current working directory (or folder)}. This means that it doesn't start with the root directory, but with the directory that you're currently in. Absolute paths are the exact location of a file or folder in the file system hierarchy. They begin with the root folder and includes all folders and subfolders necessary to locate the file or folder.}
}
}
\examples{
ger_path(".")
}
