% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_paths.R
\name{ger_path}
\alias{ger_path}
\title{Get good enough file path(s)}
\usage{
ger_path(..., ext = NULL, rel = TRUE)
}
\arguments{
\item{...}{arguments passed to \code{dir()}}

\item{ext}{optional file extension (with or without the dot (\code{.}))}

\item{rel}{logical, relative or absolute path (default is relative)}
}
\value{
file path(s) that are either normalized (\code{rel = FALSE}) or relative.
}
\description{
Get good enough file path(s)
}
\examples{
ger_path(".")
}
