% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_proj.R
\name{ger_proj}
\alias{ger_proj}
\title{Create a new good enough R project}
\usage{
ger_proj(parent_dir, proj_name)
}
\arguments{
\item{parent_dir}{the parent folder of your project}

\item{proj_name}{name of your project}
}
\description{
Create and open a new 'good enough' project.
}
\details{
this function will open a new RStudio project. The parent_dir is
    the folder that will contain your new project, and the proj_name will
    be the name of the new sub-folder.
    To navigate folders on your computer, I like to use the dir() function.
    If you start with dir("/"), you should be able to find all the folders
    you have permissions to (probably best to start with "/Users").
}
\examples{
# not run
# use dir() to find the parent folder path.
# ger_proj(parent_dir = "/Users/mjfrigaard/Documents/", "my new project")
}
