% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_report.R
\name{ger_report}
\alias{ger_report}
\title{good enough R report folder}
\usage{
ger_report(folder_name = NULL)
}
\description{
Create report folder with \code{report/manuscript.Rmd}

\describe{
\item{report/manuscript.Rmd}{A more polished R Markdown file for reports or manuscripts.}
}
}
