% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{ger_root}
\alias{ger_root}
\title{Get the root folder of your project}
\usage{
ger_root(tree = TRUE)
}
\arguments{
\item{tree}{logical, return a folder tree?}
}
\value{
path or path + folder tree
}
\description{
Get the root folder of your project
}
\examples{
ger_root()
ger_root(tree = TRUE)
}
