% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_sect.R, R/headers.R
\name{ger_sect}
\alias{ger_sect}
\title{R script section header \verb{<(+_+)>}}
\usage{
ger_sect(name)

ger_sect(name)
}
\arguments{
\item{name}{}
}
\description{
The \code{fun_headr()} function is similar to the hot keys
'cmd + shift + R' for creating a section header.

The \code{fun_headr()} function is similar to the hot keys
'cmd + shift + R' for creating a section header.
}
\examples{
# use this to create a fun section header in any .R file
# with RStudio running...
# gerp::fun_headr("import")

# use this to create a fun section header in any .R file
# with RStudio running...
# gerp::fun_headr("import")

}
