% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_setup.R
\name{ger_setup}
\alias{ger_setup}
\title{good enough R setup}
\usage{
ger_setup(folder_name = NULL)
}
\arguments{
\item{folder_name}{path to project folder}
}
\description{
Create start-up files for project (\code{changelog.md}, \code{CITATION},
\code{requirements.md}, and \code{LICENSE})

\describe{
\item{changelog.md}{Manually document changes to the files or folders in your project.}
\item{CITATION}{Information and example of how to cite your project}
\item{requirements.md}{Advice on how to manually list the requirements for your project}
\item{LICENSE}{Downloaded license from \href{https://creativecommons.org/publicdomain/zero/1.0/legalcode.txt}{creativecommons.org}}
}
}
