% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_setup.R
\name{ger_setup}
\alias{ger_setup}
\title{good enough R setup}
\usage{
ger_setup(code = TRUE, data = TRUE, docs = TRUE)
}
\arguments{
\item{code}{logical, include results code and files?}

\item{data}{logical, include data folder and files?}

\item{docs}{logical, include docs folder and files?}
}
\description{
ger_setup() is a wrapper around the ger_code(), ger_data(),
and ger_docs() functions. This is intended for an easy setup with a
new project.
}
