% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerp-package.R
\docType{package}
\name{gerp-package}
\alias{gerp}
\alias{gerp-package}
\title{gerp: Good Enough R (version 2!)}
\description{
This package helps users adopt a set of good enough computing practices when working with data. The methods are largely based on the paper, 'Good enough practices in scientific computing' by Wilson et al (2017). DOI: https://doi.org/10.1371/journal.pcbi.1005510
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mjfrigaard.github.io/gerp/}
}

}
\keyword{internal}
