% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wu_df}
\alias{wu_df}
\title{Title (dataset description)}
\format{
A dataset with variables:
\describe{
  \item{member}{Wu-Tang Clan Member Name}
  \item{name}{Wu-Tang Clan Given Name}
  \item{born}{Birth Year}
  \item{city}{Birth City}
  \item{alive}{Stil Alive?}
}
}
\source{
Link to data: https://en.wikipedia.org/wiki/Wu-Tang_Clan
(downloaded YYYY-MM-DD)
}
\usage{
wu_df
}
\description{
Title (dataset description)
}
\keyword{datasets}
